pyproject.toml
[tool.poetry]
name = "scrapy-chemical-parser"
version = "0.1.0"
description = ""
authors = ["Bohdan Salamakha <allen.avanheim@gmail.com>"]
readme = "README.md"
packages = [{include = "scrapy_chemical_parser"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
scrapy = "^2.9.0"
scrapyd = "^1.4.2"
python-dotenv = "^1.0.0"
celery = "^5.3.1"
djangorestframework = "^3.14.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.6"
flower = "^2.0.0"
redis = "^4.6.0"
scrapy-splash = "^0.9.0"
black = "^23.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
